update gemm_rs_overlap to input split
