+++ primus_turbo/pytorch/kernels/async_tp/common_ops.py
@@ -1,8 +1,19 @@
 import torch
 import triton
 import triton.language as tl
+from hip import hip
 
 import primus_turbo
+
+
+def hip_check(call_result):
+    err = call_result[0]
+    result = call_result[1:]
+    if len(result) == 1:
+        result = result[0]
+    if isinstance(err, hip.hipError_t) and err != hip.hipError_t.hipSuccess:
+        raise RuntimeError(str(err))
+    return result
 
 
 @triton.jit
